% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phea.R
\name{calculate_formula}
\alias{calculate_formula}
\title{Calculate phenotype formula(s)}
\usage{
calculate_formula(
  components,
  fml = NULL,
  window = NA,
  export = NULL,
  add_components = NULL,
  .ts = NULL,
  .pid = NULL,
  .delay = NULL,
  .line = NULL,
  .require_all = FALSE,
  .lim = NA,
  .dont_require = NULL,
  .filter = NULL,
  .cascaded = TRUE,
  .clip_sql = FALSE
)
}
\arguments{
\item{components}{A list of components, a record source, or a lazy table. If a record source or lazy table is
provided, a default component will be made from it.}

\item{fml}{Formula or list of formulas.}

\item{export}{List of additional variables to export.}

\item{add_components}{Additional components. Used mostly in case components is not a list of components.}

\item{.ts, .pid, .delay, .line}{If supplied, these will overwrite those of the given component.}

\item{.require_all}{If \code{TRUE}, returns only rows where all components to have been found according to their
timestamps (even if their value is NA). If \code{.dont_require} is provided, \code{.require_all} is ignored.}

\item{.lim}{Maximum number of rows to return. This is imposed before the calculation of the formula.}

\item{.dont_require}{If provided, causes formula to require all components (regardless of .require_all), except for
those listed here.}

\item{.filter}{Character string or list of strings. Logical conditions to satisfy. These go into the SQL \code{WHERE}
clause. Only rows satisfying all conditions provided will be returned.}

\item{.cascaded}{If \code{TRUE} (default), each formula is computed in a separate, nested SELECT statement. This allows
the result of the prior formula to be used in the following, at the potential cost of longer computation times.}

\item{.clip_sql}{If \code{TRUE}, instead of lazy table it returns the SQL query as a SQL object (can be converted to
character using \code{as.character()}), and also copies it to the clipboard.}
}
\value{
Lazy table with result of formula or formulas.
}
\description{
Receives a list of components, and a formula in SQL language, and computes the result by gathering records according
to their timestamps.
}
\details{
The data type of the columns from the components (only those that are actually used or exported) cannot be Boolean.
}
