% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_shorthands.R
\name{code_shot}
\alias{code_shot}
\title{Code shot}
\usage{
code_shot(lazy_tbl, clip = TRUE, show = TRUE)
}
\arguments{
\item{lazy_tbl}{Lazy table to look at.}

\item{clip}{If \code{TRUE} (default), will write to clipboard.}

\item{show}{If \code{FALSE} (default is \code{TRUE}), will use \code{\link[=invisible]{invisible()}} to avoid implicitly printing query to console.
This is normally used in conjunction with \code{clip = TRUE}.}
}
\value{
SQL object containing the query. Can be coerced to character by \code{as.character()}.
}
\description{
See the SQL code behind a \code{lazy_tbl}.
}
\details{
Obtain the SQL code rendered from \code{lazy_tbl}, write to the clipboard, and return it. Optionally, don't write to
clipboard.
}
