% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phea.R
\name{colnames_dbplyr}
\alias{colnames_dbplyr}
\title{\subsection{phe2a}{

PHEnotyping Algebra
By Fabrício Kury, fab at kury.dev.
2022-11-17 19:07 EST
}}
\usage{
colnames_dbplyr(records)
}
\description{
\subsection{phe2a}{

PHEnotyping Algebra
By Fabrício Kury, fab at kury.dev.
2022-11-17 19:07 EST
}
}
