% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phea.R
\docType{data}
\name{.pkgglobalenv}
\alias{.pkgglobalenv}
\title{\subsection{phea}{

PHEnotyping Algebra
By Fabrício Kury, fab at kury.dev.
2022-11-17 19:07 EST
}}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.pkgglobalenv
}
\description{
\subsection{phea}{

PHEnotyping Algebra
By Fabrício Kury, fab at kury.dev.
2022-11-17 19:07 EST
}
}
\keyword{datasets}
