% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phea.R
\name{head_shot}
\alias{head_shot}
\title{Head shot}
\usage{
head_shot(lazy_tbl, nrows = 10, blind = FALSE, .title = NA)
}
\arguments{
\item{lazy_tbl}{Lazy table to look at.}

\item{blind}{If true, will not call \code{View()}, but just return the result.}

\item{.title}{If provided, this is the title of \code{View()}. If not, \code{deparse(substitute(lazy_tbl))} will be used.}

\item{nrow}{Number of rows to collect.}
}
\value{
Collected tibble containing the first \code{nrows} of \code{lazy_tbl}.
}
\description{
Peek at the first rows of \code{lazy_tbl}.
}
\details{
Collects the first \code{nrows} of \code{lazy_tbl} into a tibble, calls \code{View()}, then returns the tibble.
}
