% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phea.R
\name{keep_row_by}
\alias{keep_row_by}
\title{Keep \link{first} row by \link{window function}}
\usage{
keep_row_by(lazy_tbl, by, partition, win_fn = "min")
}
\arguments{
\item{lazy_tbl}{Lazy table to be filtered.}

\item{by}{Column to filter rows by. Can be quoted or unquoted.}

\item{partition}{Character. Variable or vector of variables to define the partition.}

\item{win_fn}{"min" or "max". Which window function to use to filter rows.}
}
\value{
Lazy table with filtered rows.
}
\description{
Keeps the row containing the group-wise maximum or minimum.
}
\details{
Divides lazy_tbl according to \code{partition}, and in each partition keeps only the row containing the maximum or minimum
of column \code{by}.
}
