% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phea.R
\name{make_component}
\alias{make_component}
\title{Make component}
\usage{
make_component(
  input_source,
  line = NA,
  delay = NA,
  window = Inf,
  rec_name = NA
)
}
\arguments{
\item{input_source}{A record source, a component, or a lazy table.}

\item{line}{Interger. Which line to pick. 0 = skip no lines, 1 = skip one line, 2 = skip two lines, etc.}

\item{delay}{Character. Time interval in SQL language. Minimum time difference between phenotype date and component
date.}

\item{window}{Character. Time interval in SQL language. Maximum time difference between phenotype date and component
date.}

\item{rec_name}{Character. If provided, overwrites the \code{rec_name} of the record source.}
}
\value{
Phea component object.
}
\description{
Produce a Phea component.
}
\details{
Creates a component from the given \code{input_source} record source and optional parameters. If \code{input_source} is a
record source, it is used. If it is a component, it is copied (including its record source) and other paremeters, if
provided, overwrite existing ones.
}
