% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phea-package.R
\docType{package}
\name{Phea}
\alias{Phea}
\title{Phea}
\description{
Functions performing electronic patient phenotyping. Phea provides functions for building SQL queries that compute
formulas using patient data.
}
