% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phea.R
\name{sql0}
\alias{sql0}
\title{SQL query}
\usage{
sql0(...)
}
\arguments{
\item{...}{Arguments to be coerced to \code{character} and concatenated.}
}
\value{
Lazy table corresponding to the query.
}
\description{
Combines with \code{paste0} the strings in \code{...}, then runs it as a SQL query.
}
