% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_shorthands.R
\name{sqla}
\alias{sqla}
\title{SQL query with arguments}
\usage{
sqla(args, ..., no_check = FALSE)
}
\arguments{
\item{args}{}

\item{...}{Character strings to be concatenated with \code{paste0}.}
}
\value{
Lazy table corresponding to the query.
}
\description{
Combines with \code{paste0} the strings in \code{...}, then runs it as a SQL query using \code{args} as common table expressions.
The names of the \code{args} objects are the names made available in the query.
}
\examples{
```
list(a = sqlt(person), b = sqlt(procedure)) |>
  sqla('select person_id fr', 'om a inner jo', 'in b on a.person_id = b.person_id')
```
}
\seealso{
\code{\link[=sqlt]{sqlt()}} to create lazy tables from SQL tables. \code{\link[=sql0]{sql0()}} to run arbitrary SQL.
}
