% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phea.R
\name{keep_row_by}
\alias{keep_row_by}
\title{Keep \link{first or last} row by \link{window function}}
\usage{
keep_row_by(
  lazy_tbl,
  by,
  partition,
  pick_last = FALSE,
  .fn = NULL,
  .fn_arg = NULL,
  .val = NULL
)
}
\arguments{
\item{lazy_tbl}{Lazy table to be filtered.}

\item{by}{Column to pick rows by.}

\item{partition}{Character vector. Variable or variables to define the partition.}

\item{pick_last}{Logical. If \code{TRUE}, will pick the last row, instead of first.}
}
\value{
Lazy table with filtered rows.
}
\description{
Keeps the row containing the group-wise maximum or minimum.
}
\details{
Divides lazy_tbl according to \code{partition}, and in each partition keeps only the row containing the maximum or minimum
of column \code{by}.
}
