% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phea.R
\name{make_component}
\alias{make_component}
\title{Make component}
\usage{
make_component(
  input_source,
  line = NA,
  delay = NA,
  window = Inf,
  rec_name = NA,
  .passthrough = FALSE,
  .ts = NULL,
  .pid = NULL
)
}
\arguments{
\item{input_source}{A record source from \code{make_record_source()}, a component from \code{make_component()}, or a lazy
table. If the latter case, \code{.ts} and \code{.pid} must be provided.}

\item{line}{Interger. Which line to pick. 0 = skip no lines, 1 = skip one line, 2 = skip two lines, etc.}

\item{delay}{Character. Time interval in SQL language. Minimum time difference between phenotype date and component
date.}

\item{window}{Character. Time interval in SQL language. Maximum time difference between phenotype date and component
date.}

\item{.ts}{Unquoted character. If passing a lazy table to \code{input_source}, \code{.ts} is used as \code{ts} to buid a record
source.}

\item{.pid}{Unquoted character. If passing a lazy table to \code{input_source}, \code{.pid} is used as \code{pid} to buid a record
source.}
}
\value{
Phea component object.
}
\description{
Produce a Phea component.
}
\details{
Creates a component from the given \code{input_source} record source and optional parameters. If \code{input_source} is a
record source, it is used. If it is a component, it is copied (including its record source) and other paremeters, if
provided, overwrite existing ones.
}
