% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phea.R
\name{make_record_source}
\alias{make_record_source}
\title{Make record source}
\usage{
make_record_source(
  records,
  ts,
  pid,
  rec_name = NULL,
  vars = NULL,
  .ts = NULL,
  .pid = NULL
)
}
\arguments{
\item{records}{Lazy table with records to be used.}

\item{ts}{Unquoted characters. Name of the colum in \code{records} that gives the timestamp.}

\item{pid}{Unquoted characters. Name of the colum in \code{records} that gives the person (patient) identifier.}

\item{rec_name}{Character. Optional. Record name.}

\item{vars}{Character vector. Optional. Name of the colums to make available from \code{records}. If not supplied, all
columns are used.}
}
\value{
Phea record source object.
}
\description{
Create a Phea record source.
}
\details{
Creates a record source from a lazy table.
}
\seealso{
\code{\link[=make_component]{make_component()}} to create a component from a record source.
}
