% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phea.R
\name{make_record_source}
\alias{make_record_source}
\title{Make record source}
\usage{
make_record_source(
  records,
  rec_name = NULL,
  ts,
  pid,
  vars = NULL,
  .capture_col = NULL,
  .type = "direct",
  .ts = NULL,
  .pid = NULL
)
}
\arguments{
\item{records}{Lazy table with records to use.}

\item{rec_name}{Character. Record name.}

\item{ts}{Unquoted string. Name of the colum in \code{records} that gives the timestamp.}

\item{pid}{Unquoted string. Name of the colum in \code{records} that gives the person (patient) identifier.}

\item{vars}{Character vector. Name of the colums to make available from \code{records}. If not supplied, all columns are
used.}

\item{.capture_col}{Unquoted string. Not yet implemented.}
}
\value{
Phea record source object.
}
\description{
Create a Phea record source.
}
\details{
Creates a record source from a lazy table.
}
