% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phea.R
\name{make_record_source}
\alias{make_record_source}
\title{make_record_source}
\usage{
make_record_source(
  records,
  rec_name,
  ts,
  pid,
  vars = NULL,
  .capture_col = NULL
)
}
\arguments{
\item{records}{List of records.}

\item{rec_name}{Record name.}
}
\value{
Record source..
}
\description{
make_record_source
}
