% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{phea_plot}
\alias{phea_plot}
\title{Plot a phenotype.}
\usage{
phea_plot(
  board,
  pid,
  plot_title = NULL,
  exclude = NULL,
  verbose = NULL,
  .board = NULL
)
}
\arguments{
\item{board}{Phenotype object returned by \code{calculate_formula()}.}

\item{pid}{Required. ID of the patient to be included in the chart.}

\item{exclude}{Optional. Names of columns to not plot.}

\item{verbose}{If TRUE, will let you know how long it takes to \code{collect()} the data. If not provided, defaults to
\code{.verbose} provided to \code{setup_phea()}, which itself defaults to \code{TRUE} if not provided.}

\item{.board}{Optional. Local data frame. Provide this argument together with \code{board = NULL} to use local data
directly, instead of \code{collect()}ing the \code{board}.}
}
\value{
Plot created by \code{plotly::plot_ly()} within \code{plotly::subplot()}.
}
\description{
Plots the result from \code{calculate_formula()} in an interactive timeline chart using \code{plotly}.
}
\details{
Collects (downloads) the results and creates interactive timeline chart using the \code{plotly} library.
}
