% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phea.R
\name{pick_row_by}
\alias{pick_row_by}
\title{Pick \link{first or last} row by \link{window function}}
\usage{
pick_row_by(
  lazy_tbl,
  by,
  partition,
  pick_last = FALSE,
  .fn = NULL,
  .fn_arg = NULL,
  .val = NULL
)
}
\arguments{
\item{lazy_tbl}{Lazy table to be filtered.}

\item{by}{Character. Column(s) to pick rows by.}

\item{partition}{Character vector. Column name(s) to define the partition.}

\item{pick_last}{Logical. If \code{TRUE}, will pick the last row, instead of first.}

\item{.fn}{Character. Name of the aggregate function to use, \emph{without parentheses}. E.g.: \code{max}, \code{cume_dist}.}

\item{.val}{Character or numeric. Literal value to compare to result of \code{.fn}.}
}
\value{
Lazy table with filtered rows.
}
\description{
Pick the rows that contain the group-wise aggregate value in each partition.
}
\details{
Divides \code{lazy_tbl} according to \code{partition}, and in each one keeps only the row picked by the result of a window
function \code{.fn}.

If \code{.fn} is not provided, defaults to picking the rows where \code{by} is maximum.

If \code{pick_last} is \code{TRUE}, defaults instead to rows where \code{by} is minimum.

If \code{.val} is provided, keeps only the rows where the result of \code{.fn(by)} in each partition is equal to \code{.val}.
}
