% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_change.R
\name{pick_row_by}
\alias{pick_row_by}
\title{Pick row by window function}
\usage{
pick_row_by(
  lazy_tbl,
  by,
  partition,
  pick_max = FALSE,
  .fn = NULL,
  .fn_arg = NULL,
  .val = NULL
)
}
\arguments{
\item{lazy_tbl}{Lazy table to be filtered.}

\item{by}{Character. Column(s) to pick rows by.}

\item{partition}{Character vector. Column name(s) to define the partition.}

\item{pick_max}{Logical. If \code{TRUE}, will pick the last row, instead of first.}

\item{.fn}{Character. Name of the aggregate function to use, \emph{without parentheses}. E.g.: \code{max}, \code{cume_dist}.}

\item{.val}{Character or numeric. Literal value to compare to result of \code{.fn}.}
}
\value{
Lazy table with filtered rows.
}
\description{
Pick (keep) the row(s) that contain the group-wise aggregate value of \code{.fn} in each partition. If \code{.fn} is not
provided, defaults to picking the rows where the value of \code{by} is minimum.
}
\details{
If \code{pick_min} is \code{TRUE}, defaults instead to rows where \code{by} is maximum.

If \code{.val} is provided, keeps the rows where the result of \code{.fn(by)} is equal to \code{.val}.
}
