% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phea.R
\name{setup_phea}
\alias{setup_phea}
\title{Setup Phea}
\usage{
setup_phea(connection, schema)
}
\arguments{
\item{connection}{DBI-compatible SQL connection (e.g. produced by DBI::dbConnect).}

\item{schema}{Schema to be used by default in \code{sqlt()}.}
}
\description{
Configures functions \code{sqlt()} and \code{sql0()} for use.
}
\details{
Stores the DBI connection for later use.
}
