% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_phea}
\alias{setup_phea}
\title{Setup Phea}
\usage{
setup_phea(connection, schema, verbose = TRUE, .fix_dbplyr_spark = FALSE)
}
\arguments{
\item{connection}{DBI-compatible SQL connection (e.g. produced by DBI::dbConnect).}

\item{schema}{Schema to be used by default in \code{sqlt()}. If no schema, use \code{NA}.}

\item{verbose}{Logical. Optional. If TRUE (default), functions will print to console at times.}

\item{.fix_dbplyr_spark}{Logical. Optional. Very niche functionality. Set to \code{TRUE} to attempt to fix the use of
\verb{IGNORE NULLS} by the OBDC driver connected to a Spark SQL server/cluster.}
}
\description{
Configures Phea, in particular the SQL shorthands \code{sqlt()}, \code{sql0()} and \code{sqla()}.
}
