% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phea.R
\name{sql0}
\alias{sql0}
\title{SQL query}
\usage{
sql0(..., schema = NULL)
}
\arguments{
\item{...}{Character strings to be concatenated with \code{paste0}.}

\item{schema}{Optional. Name of schema to use. If not provided, will default to schema passed to \code{setup_phea()}.}
}
\value{
Lazy table corresponding to the query.
}
\description{
Combines with \code{paste0} the strings in \code{...}, then runs it as a SQL query.
}
\seealso{
\code{\link[=sqlt]{sqlt()}} to create lazy tables from SQL tables. \code{\link[=sqla]{sqla()}} to run arbitrary SQL with lazy tables.
}
