% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phea.R
\name{sqla}
\alias{sqla}
\title{SQL query with arguments}
\usage{
sqla(args, ...)
}
\arguments{
\item{args}{}

\item{...}{Character strings to be concatenated with \code{paste0}.}
}
\value{
Lazy table corresponding to the query.
}
\description{
Combines with \code{paste0} the strings in \code{...}, then runs it as a SQL query using \code{args} as common table expressions.
The names of the \code{args} objects are the names made available in the query.
}
\examples{
```
list(a = sqlt(person), b = sqlt(procedure)) |>
  sqla('select person_id ', 'from a ', 'inner join b on a.person_id = b.person_id')
```
}
