% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phea.R
\name{sqlt}
\alias{sqlt}
\title{SQL table}
\usage{
sqlt(table)
}
\arguments{
\item{dbi_connection}{DBI-compatible SQL connection (e.g. produced by DBI::dbConnect).}

\item{schema}{Schema to be used by default in \code{sqlt()}.}
}
\value{
Lazy table equal to \verb{select * from def_schema.table;}.
}
\description{
Produces lazy table object of table \code{table} in the preconfigured schema.
}
\details{
This function is a shorthand for \verb{select * from def_schema.table;}. \code{def_schema} is the schema passed to
\code{setup_phea()} previously.
}
