% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phea.R
\name{sqlt}
\alias{sqlt}
\title{SQL table}
\usage{
sqlt(table, schema = NULL)
}
\arguments{
\item{table}{Unquoted name of the table to be accessed within \code{def_schema}.}

\item{schema}{Optional. Name of schema to use. If not provided, will default to schema passed to \code{setup_phea()}.}
}
\value{
Lazy table equal to \verb{select * from def_schema.table;}.
}
\description{
Produces lazy table object of table \code{table} in the preconfigured schema.
}
\details{
This function is a shorthand for \verb{select * from def_schema.table;}. \code{def_schema} is the schema passed to
\code{setup_phea()} previously.
}
\examples{
`sqlt(person)`
`sqlt(condition_occurrence)`
}
